requirements.txt
langchain
langchain-google-genai
pypdf
chromadb
python-dotenv
langchain-community
streamlit